+++ src/back/Action/TopicControl.php
@@ -8,11 +8,11 @@ use KeepersTeam\Webtlo\Clients\ClientFactory;
 use KeepersTeam\Webtlo\Clients\ClientInterface;
 use KeepersTeam\Webtlo\Clients\Data\Torrents;
 use KeepersTeam\Webtlo\Config\TopicControl as ConfigControl;
-use KeepersTeam\Webtlo\DTO\KeysObject;
 use KeepersTeam\Webtlo\Module\Control\ApiSearch;
 use KeepersTeam\Webtlo\Module\Control\DbSearch;
 use KeepersTeam\Webtlo\Module\Control\PeerCalc;
 use KeepersTeam\Webtlo\Module\Control\Unseeded;
+use KeepersTeam\Webtlo\Storage\KeysObject;
 use KeepersTeam\Webtlo\Timers;
 use Psr\Log\LoggerInterface;
 use RuntimeException;
