+++ src/back/Module/Report/CreateReport.php
@@ -10,11 +10,11 @@ use KeepersTeam\Webtlo\Config\Credentials;
 use KeepersTeam\Webtlo\Config\ReportSend;
 use KeepersTeam\Webtlo\Data\Forum;
 use KeepersTeam\Webtlo\DB;
-use KeepersTeam\Webtlo\DTO\KeysObject;
 use KeepersTeam\Webtlo\Enum\UpdateMark;
 use KeepersTeam\Webtlo\Enum\UpdateStatus;
 use KeepersTeam\Webtlo\Helper;
 use KeepersTeam\Webtlo\Settings;
+use KeepersTeam\Webtlo\Storage\KeysObject;
 use KeepersTeam\Webtlo\Storage\Table\Forums;
 use KeepersTeam\Webtlo\Storage\Table\UpdateTime;
 use KeepersTeam\Webtlo\WebTLO;
